Extract JsonLd RDFTerm types
